64-bit patches
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1003 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/include/G3D/Array.h
@@ -740,7 +740,7 @@ public:
        sequence, a value at least as large as size()" 
        For compatibility with std::vector.
    */
-   int capacity() const {
+   size_t capacity() const {
        return numAllocated;
    }
